fix: preserve fractional Waymark balances

+++ libraries/wayfarer-api/src/main/java/io/github/eariver/wayfarer/api/WayfarerWaymark.java
@@ -1,10 +1,14 @@
 package io.github.eariver.wayfarer.api;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import java.util.concurrent.CompletionStage;
 
 public interface WayfarerWaymark {
-    CompletionStage<Long> balance(UUID playerUuid);
+    /**
+     * Reads the provider-authoritative balance without rounding it to an integer.
+     */
+    CompletionStage<BigDecimal> balance(UUID playerUuid);
     CompletionStage<OperationResult> debit(UUID playerUuid, long amount, String reference);
     CompletionStage<OperationResult> credit(UUID playerUuid, long amount, String reference);
     default CompletionStage<OperationResult> refund(
